Added ==smoke-test support and ciftest command.
Commands can now define a "==smoke-test" section in their CIF files, which defines a snippet of fshell script that will be run as part of "fshell smoketest" or by invoking ciftest directly. See the ciftest documentation for more details.

Added ==smoke-test sections to a few commands, the ones that were easy to test!